Feat/mip custom novel2 (#1290)
* feat: 广告合并-bugfix #314

* feat: 广告合并-bugfix #314

* feat: 广告合并-bugfix

* feat: 广告合并-bugfix

* feat: 针对小说添加新的判断条件，并且把小说的内容从custom文件中抽离出来 #1289

* feat: 针对小说添加新的判断条件，并且把小说的内容从custom文件中抽离出来 #1289

* feat: 针对小说添加新的判断条件，并且把小说的内容从custom文件中抽离出来 #1289
